+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/Lib/Helpers.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.IO;
+using System.Text;
 using System.Security.Cryptography;
 using Renci.SshNet;
 using Renci.SshNet.Common;
+using Renci.SshNet.Security;
 using Frends.SFTP.UploadFiles.Definitions;
 
 namespace Frends.SFTP.UploadFiles.Tests;
@@ -15,6 +17,7 @@ internal static class Helpers
     readonly static string _dockerAddress = "localhost";
     readonly static string _dockerUsername = "foo";
     readonly static string _dockerPassword = "pass";
+    readonly static string _workDir = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../TestData/");
 
     internal static Connection GetSftpConnection()
     {
@@ -31,14 +34,6 @@ internal static class Helpers
         };
 
         return connection;
-    }
-
-    internal static void DeleteSourceDir()
-    {
-        var year = DateTime.Now.Year.ToString();
-        var dir = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../TestData/testfolder_" + year);
-        if (Directory.Exists(dir))
-            Directory.Delete(dir, true);
     }
 
     internal static string GetServerFingerprintAsSHA256String()
@@ -140,6 +135,79 @@ internal static class Helpers
             }
             sftp.Disconnect();
         }
+    }
+
+    internal static void CreateDummyFiles(int count)
+    {
+        var name = "SFTPUploadTestFile";
+        var extension = ".txt";
+        for (var i = 1; i <= count; i++)
+        {
+            File.WriteAllText(Path.Combine(_workDir, name + i + extension), "This is a test file.");
+        }
+    }
+
+    internal static void DeleteDummyFiles()
+    {
+        foreach (var file in Directory.EnumerateFileSystemEntries(_workDir))
+        {
+            if (Directory.Exists(file))
+                Directory.Delete(file, true);
+            if (!file.Contains("LargeTestFile.bin"))
+                File.Delete(file);
+        }
+    }
+
+    internal static Tuple<byte[], byte[]> GetServerFingerPrintAndHostKey()
+    {
+        Tuple<byte[], byte[]> result = null;
+        using (var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
+        {
+            client.ConnectionInfo.HostKeyAlgorithms.Clear();
+            client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
+
+            client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
+            {
+                result = new Tuple<byte[], byte[]>(e.FingerPrint, e.HostKey);
+                e.CanTrust = true;
+            };
+            client.Connect();
+            client.Disconnect();
+        }
+        return result;
+    }
+
+    internal static string ConvertToMD5Hex(byte[] fingerPrint)
+    {
+        return BitConverter.ToString(fingerPrint).Replace("-", ":");
+    }
+
+    internal static string ConvertToSHA256Hash(byte[] hostKey)
+    {
+        var fingerprint = "";
+        using (SHA256 mySHA256 = SHA256.Create())
+        {
+            fingerprint = Convert.ToBase64String(mySHA256.ComputeHash(hostKey));
+        }
+        return fingerprint;
+    }
+
+    internal static string ConvertToSHA256Hex(byte[] hostKey)
+    {
+        var fingerprint = "";
+        using (SHA256 mySHA256 = SHA256.Create())
+        {
+            fingerprint = ToHex(mySHA256.ComputeHash(hostKey));
+        }
+        return fingerprint;
+    }
+
+    internal static string ToHex(byte[] bytes)
+    {
+        StringBuilder result = new StringBuilder(bytes.Length * 2);
+        for (int i = 0; i < bytes.Length; i++)
+            result.Append(bytes[i].ToString("x2"));
+        return result.ToString();
     }
 }