+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/MacroTests.cs
@@ -24,7 +24,7 @@ class MacroTests : UploadFilesTestBase
         var result = SFTP.UploadFiles(_source, destination, _connection, _options, _info, new CancellationToken());
         Assert.IsTrue(result.Success);
         var date = DateTime.Now;
-        Assert.IsTrue(Helpers.CheckFileExistsInDestination("/upload/Upload/SFTPUploadTestFile" + date.ToString(@"yyyy-MM-dd") + ".txt"));
+        Assert.IsTrue(Helpers.CheckFileExistsInDestination("/upload/Upload/SFTPUploadTestFile1" + date.ToString(@"yyyy-MM-dd") + ".txt"));
     }
 
     [Test]