+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/Connection.cs
@@ -63,8 +63,10 @@ public class Connection
     public string PrivateKeyFile { get; set; }
 
     /// <summary>
-    /// Private key as a string, supported private key formats: PKCS#8,
-    /// PuTTY.ppk.
+    /// Private key as a string, supported private key formats: OpenSSH.
+    /// PuTTY keys can be converted with puttygen.exe application.
+    /// 1. Load your key file into puttygen.exe
+    /// 2. Conversion > Export OpenSSH key (not the "force new file format" option)
     /// </summary>
     /// <example>
     /// -----BEGIN RSA PRIVATE KEY-----