+++ Frends.SFTP.UploadFiles/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+#[2.0.1] - 2022-08-26
+### Fixed
+- Fixed issue with server fingerprint given by user in SHA256 hex format was not accepted: Added conversion to the fingerprint given by user.
+- Fixed issue when using invalid server fingerprint in MD5 string format throws wrong error message: Added more specific error messages.
+- Changed how MD5 string is handled. MD5 can now be given without ':' or '-' characters.
+- Fixed issue that Sha256 was only accepted in Base64 format: Added support for Sha256 in hex format.
+- Changed the used HostKeyAlgorithm by forcing to use ssh-rsa default was ed25519.
+- Added more tests for using server fingerprints.
+- Fixed issue where when using SourceOperation.Move the source file cannot be restored when exception occurs. 
+- Removed FileEncoding UTF-16 which was not implemented and threw exception if selected.
+
 # [2.0.0] - 2022-08-08
 ### Fixed
 - [Breaking] Changed the implementation to work similar to Cobalt by moving the source file to local Temp folder before transferring to destination.